New Configuration Field: videoBitrateInBps: Long? for granular bitrate control BPS takes precedence over Mbps when both are specified Helper Methods:     - Configuration.withBitrateInBps() - Creates config with bps bitrate     - Configuration.withBitrateInMbps() - Creates config with Mbps bitrate (legacy)     - getEffectiveBitrateInBps() - Internal method to resolve bitrate

+++ lightcompressor/src/main/java/com/abedelazizshe/lightcompressorlibrary/utils/CompressorUtils.kt
@@ -69,7 +69,7 @@ object CompressorUtils {
     fun setOutputFileParameters(
         inputFormat: MediaFormat,
         outputFormat: MediaFormat,
-        newBitrate: Int,
+        newBitrate: Long,
     ) {
         val newFrameRate = getFrameRate(inputFormat)
         val iFrameInterval = getIFrameIntervalRate(inputFormat)
@@ -93,7 +93,7 @@ object CompressorUtils {
             setInteger(MediaFormat.KEY_FRAME_RATE, newFrameRate)
             setInteger(MediaFormat.KEY_I_FRAME_INTERVAL, iFrameInterval)
             // expected bps
-            setInteger(MediaFormat.KEY_BIT_RATE, newBitrate)
+            setInteger(MediaFormat.KEY_BIT_RATE, newBitrate.toInt())
             setInteger(
                 MediaFormat.KEY_BITRATE_MODE,
                 MediaCodecInfo.EncoderCapabilities.BITRATE_MODE_CBR
